+++ src/main/java/coupon/coupon/domain/Coupon.java
@@ -12,7 +12,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 @Entity
 public class Coupon {
