+++ src/main/java/coupon/coupon/domain/DiscountAmount.java
@@ -5,7 +5,7 @@ import java.math.RoundingMode;
 import java.util.Objects;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 @Embeddable
 public class DiscountAmount {
