+++ src/main/java/coupon/coupon/domain/MinimumOrderAmount.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.Objects;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 @Embeddable
 public class MinimumOrderAmount {
