+++ src/main/java/coupon/coupon/domain/Term.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import java.util.Objects;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 @Embeddable
 public class Term {
