+++ src/main/java/coupon/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
-package coupon.coupon.repository;
+package coupon.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import coupon.coupon.domain.Member;
+import coupon.member.domain.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 }
