+++ src/test/java/coupon/coupon/domain/CouponNameTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
