+++ src/test/java/coupon/coupon/domain/DiscountAmountTest.java
@@ -3,7 +3,7 @@ package coupon.coupon.domain;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
