+++ src/test/java/coupon/member/domain/NicknameTest.java
@@ -1,10 +1,10 @@
-package coupon.coupon.domain;
+package coupon.member.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import coupon.coupon.CouponException;
+import coupon.CouponException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
